+++ src/components/face-challenge/facetime/CustomVideoConference.tsx
@@ -1,39 +1,9 @@
 import { TrackLoop, useTracks } from '@livekit/components-react';
 import { Track } from 'livekit-client';
-import React, { useCallback } from 'react';
-import { useFacetimeChallengeStore } from '../../store/facetimeChallengeStore';
-import Button from '../common/Button';
+import React from 'react';
+import useChallengeLogic from '../../../hooks/useChallengeLogin';
+import Button from '../../common/Button';
 import CustomParticipantTile from './CustomParticipantTile';
-
-const useChallengeLogic = () => {
-  const { status, setStatus } = useFacetimeChallengeStore();
-  const camaraTracks = useTracks(
-    [{ source: Track.Source.Camera, withPlaceholder: true }],
-    { onlySubscribed: false },
-  );
-  const handleStartChallenge = useCallback(() => {
-    if (status !== 'idle') return;
-    if (camaraTracks.length < 2) {
-      alert('가족들이 모두 참여해야 합니다.');
-      return;
-    }
-    setStatus('ongoing');
-  }, [status, camaraTracks, setStatus]);
-
-  const handleChangeStatus = useCallback(() => {
-    if (status === 'idle') {
-      handleStartChallenge();
-      return;
-    }
-
-    if (status === 'ongoing') {
-      setStatus('finished');
-      return;
-    }
-  }, [status, handleStartChallenge, setStatus]);
-
-  return { handleChangeStatus, camaraTracks, status };
-};
 
 const CustomVideoConference = () => {
   const camaraTracks = useTracks(
