+++ src/components/face-challenge/facetime/FacetimeContainer.tsx
@@ -48,7 +48,6 @@ const FacetimeContainer = ({ code }: FacetimeContainerProps) => {
         }}
         variation="minimal"
       />
-
       <FilterSelector />
       <Button
         onClick={() => {
