+++ src/components/face-challenge/facetime/OtherParticipantTile.tsx
@@ -5,8 +5,8 @@ import {
   useTrackRefContext,
 } from '@livekit/components-react';
 import { useEffect, useState } from 'react';
-import useFaceFilter from '../../hooks/useFaceFilter';
-import { FilterPosition, FilterType } from '../../types/challenge';
+import useFaceFilter from '../../../hooks/useFaceFilter';
+import { FilterPosition, FilterType } from '../../../types/challenge';
 
 const decoder = new TextDecoder();
 const OtherParticipantTile = () => {
